Refactor Kafka Streaming table into datafusion/core
